pyproject.toml
[project]
name = "lending-erc20-protocol"
description = "Lending Protocol by Zharta"
classifiers = ["Private :: Do Not Upload"]
version = "0"
dependencies = [
    "eth-ape",
    "vyper==0.4.3",
    "ape-vyper",
    "ape-foundry",
    "ape-alchemy",
    "ape-arbitrum",
    "ape-base",
    "web3",
]


[project.optional-dependencies]
dev = [
    "titanoboa==0.2.7",
    "boto3",
    "click",
    "coverage",
    "hypothesis",
    "ipython",
    "mypy",
    "pre-commit",
    "pytest",
    "pytest-bdd",
    "pytest-xdist",
    "python-lsp-server",
    "rich",
    "rope",
    "ruff",
    # "vyper-lsp",
]

[build-system]
requires = ["setuptools>=63", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["scripts"]

[tool.pytest.ini_options]
log_file = "pytest-logs.txt"
addopts = """
    -p no:ape_test
    -vv
    --durations=10
"""

[tool.ruff]
lint.select = ["ALL"]
lint.ignore = ["ANN", "B905", "BLE", "COM812", "CPY", "D", "DTZ", "EM", "FIX", "ISC001", "PLR0913", "PLR2004", "S", "TCH", "TD", "TRY003"]
line-length = 127
lint.preview = true
target-version = "py311"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"ape_console_extras.py" = ["T201", "B006", "PYI024", "RUF052"]
"tests/*.py" = [
    "ARG001",
    "DOC",
    "ERA001",
    "F401",
    "FBT003",
    "FURB118",
    "FURB140",
    "N806",
    "N815",
    "PLC1901",
    "PLR0914",
    "PLR0915",
    "PLR0917",
    "PT004",
    "PT022",
    "PTH",
    "PYI024",
    "RUF029",
    "RUF052",
    "RUF100",
    "SLF001",
    "T201",
    "TID252",
]
"scripts/*.py" = [
    "A001",
    "A004",
    "DOC",
    "ERA001",
    "FURB101",
    "FURB103",
    "ERA001",
    "PLW1514",
    "PTH123",
    "RUF052",
    "T201",
    "UP015"
]

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false


[tool.coverage.run]
plugins = [ "boa.coverage" ]
relative_files = true
omit = [
    "contracts/auxiliary/*",
    "tests/stubs/*",
    "None"
    ]

[tool.coverage.paths]
source = ["contracts"]

[tool.coverage.report]
show_missing = true
